split t_command constructor into one file per command type
